+++ internal/reports/sqs/parsers/types.go
@@ -6,11 +6,13 @@
 // See the LICENSE file in the root of this repository for full license text or
 // visit: <https://www.gnu.org/licenses/gpl-3.0.html>.
 
-package httpserver
+package parsers
 
-// ScheduleArtifactAnalysisRequest is the schema
-// for the POST /
-type ScheduleArtifactAnalysisRequest struct {
-	ImageURI  string `json:"imageURI"`
-	Namespace string `json:"namespace"`
-}
+import (
+	"context"
+
+	sqstypes "github.com/aws/aws-sdk-go-v2/service/sqs/types"
+	"github.com/crashappsec/chalkular/api/v1beta1/chalk"
+)
+
+type ChalkReportParser = func(context.Context, sqstypes.Message) ([]chalk.Report, error)